Add VW, split integration tests over files
